src/requirements.txt
SQLAlchemy_serializer==1.3.4.4
SQLAlchemy==1.3.20
SQLAlchemy_Utils==0.36.8
PyYAML==5.3.1
#aws-psycopg2==1.2.1
psycopg2-binary==2.8.6
python-jose==3.2.0
